+++ account/package.json
@@ -6,7 +6,7 @@
   "scripts": {
     "start:dev": "nodemon --ext js,json,yaml src/main.js",
     "start": "node src/main.js",
-    "test": "echo \"Error: no test specified\" && exit 1"
+    "test": "node --experimental-vm-modules node_modules/jest/bin/jest.js --detectOpenHandles"
   },
   "author": "Vanderleia Souza",
   "license": "ISC",
@@ -15,7 +15,9 @@
   },
   "description": "",
   "devDependencies": {
-    "nodemon": "^2.0.20"
+    "jest": "^29.3.1",
+    "nodemon": "^2.0.20",
+    "supertest": "^6.3.3"
   },
   "dependencies": {
     "bcryptjs": "^2.4.3",
