--- account/test/createUserAccount.test.js
@@ -1,5 +1,0 @@
-import { createUserUseCase } from "../src/use-case/createUserAccount.js";
-
-const user = await createUserUseCase("julia", "julia@email.com", "senhaDaJulia");
-
- console.log(user);
