+++ product/config/config.json
@@ -7,9 +7,9 @@
     "dialect": "mysql"
   },
   "test": {
-    "username": "root",
-    "password": null,
-    "database": "database_test",
+    "username": "vanderleia",
+    "password": "23456",
+    "database": "products",
     "host": "127.0.0.1",
     "dialect": "mysql"
   },
