+++ product/src/test/data/products.js
@@ -8,5 +8,5 @@ export const productExample = {
     description: "Tenis preto cocadarço",
     category: "Calçados",
     fetures: [{name: "cor", description: "preto"}],
-    images: [{url: "https://secure-static.vans.com.br/medias/sys_master/vans/vans/hb8/hf3/h00/h00/9572416782366/1002001070008U-01-BASEIMAGE-Hires.jpg", descriptions: "Tênis preto"}]
+    images: [{url: "https://secure-static.vans.com.br/medias/sys_master/vans/vans/hb8/hf3/h00/h00/9572416782366/1002001070008U-01-BASEIMAGE-Hires.jpg", description: "Tênis preto"}]
   };
